reset GlobalGrid to original settings;     no reference to paritioner: Base is abstract, specialization for dim/manualBlock partitioner needs too many changes     addition class GlobalSparseGrid for sparse grids
