+++ cajita/src/Cajita_GlobalGrid.hpp
@@ -13,12 +13,16 @@
   \file Cajita_GlobalGrid.hpp
   \brief Global grid
 */
-#ifndef CAJITA_GLOBALGRID_HPP
-#define CAJITA_GLOBALGRID_HPP
+#ifndef CAJITA_GLOBALGRID_NEW_HPP
+#define CAJITA_GLOBALGRID_NEW_HPP
 
 #include <Cajita_GlobalMesh.hpp>
-#include <Cajita_Partitioner.hpp>
+#include <Cajita_SparseIndexSpace.hpp>
 #include <Cajita_Types.hpp>
+
+#include <Cajita_Partitioner.hpp>
+#include <Cajita_SparseDimPartitioner.hpp>
+#include <Cajita_UniformDimPartitioner.hpp>
 
 #include <array>
 #include <memory>
@@ -29,11 +33,11 @@ namespace Cajita
 {
 //---------------------------------------------------------------------------//
 /*!
-  \brief Global logical grid.
+  \brief Global logical grid base, store global mesh, rank related information.
   \tparam MeshType Mesh type (uniform, non-uniform, sparse)
 */
 template <class MeshType>
-class GlobalGrid
+class GlobalGridBase
 {
   public:
     //! Mesh type.
@@ -47,15 +51,14 @@ class GlobalGrid
      \param comm The communicator over which to define the grid.
      \param global_mesh The global mesh data.
      \param periodic Whether each logical dimension is periodic.
-     \param partitioner The grid partitioner.
     */
-    GlobalGrid( MPI_Comm comm,
-                const std::shared_ptr<GlobalMesh<MeshType>>& global_mesh,
-                const std::array<bool, num_space_dim>& periodic,
-                const BlockPartitioner<num_space_dim>& partitioner );
+    GlobalGridBase( MPI_Comm comm,
+                    const std::shared_ptr<GlobalMesh<MeshType>>& global_mesh,
+                    const std::array<bool, num_space_dim>& periodic,
+                    const BlockPartitioner<num_space_dim>& partitioner );
 
     // Destructor.
-    ~GlobalGrid();
+    ~GlobalGridBase();
 
     //! \brief Get the communicator. This communicator was generated with a
     //! Cartesian topology.
@@ -93,7 +96,6 @@ class GlobalGrid
       \brief Get the MPI rank of a block with the given indices. If the rank is
       out of bounds and the boundary is not periodic, return -1 to indicate an
       invalid rank.
-
       \param ijk %Array of block indices.
     */
     int blockRank( const std::array<int, num_space_dim>& ijk ) const;
@@ -102,7 +104,6 @@ class GlobalGrid
       \brief Get the MPI rank of a block with the given indices. If the rank is
       out of bounds and the boundary is not periodic, return -1 to indicate an
       invalid rank.
-
       \param i,j,k Block index.
     */
     template <std::size_t NSD = num_space_dim>
@@ -113,7 +114,6 @@ class GlobalGrid
       \brief Get the MPI rank of a block with the given indices. If the rank is
       out of bounds and the boundary is not periodic, return -1 to indicate an
       invalid rank.
-
       \param i,j Block index.
     */
     template <std::size_t NSD = num_space_dim>
@@ -153,8 +153,50 @@ class GlobalGrid
     std::enable_if_t<3 == NSD, int> globalNumEntity( Edge<Dim::K>,
                                                      const int dim ) const;
 
-    //! \brief Get the owned number of cells in a given dimension of this block.
-    //! \param dim Spatial dimension.
+  private:
+    MPI_Comm _cart_comm;
+    std::shared_ptr<GlobalMesh<MeshType>> _global_mesh;
+    std::array<bool, num_space_dim> _periodic;
+    std::array<int, num_space_dim> _ranks_per_dim;
+    std::array<int, num_space_dim> _cart_rank;
+    std::array<bool, num_space_dim> _boundary_lo;
+    std::array<bool, num_space_dim> _boundary_hi;
+};
+
+//---------------------------------------------------------------------------//
+/*!
+  \brief Global logical grid for uniform and non-uniform grids
+  \tparam MeshType Mesh type (uniform, non-uniform)
+*/
+// template <class MeshType, class PartitionerType>
+template <class MeshType>
+class GlobalGrid : public GlobalGridBase<MeshType>
+{
+  public:
+    //! Mesh type.
+    using mesh_type = MeshType;
+
+    //! Spatial dimension
+    using GlobalGridBase<mesh_type>::num_space_dim;
+
+    //! partitioner type
+    // using partitioner_type = PartitionerType;
+
+    /*!
+       \brief Constructor.
+       \param comm The communicator over which to define the grid.
+       \param global_mesh The global mesh data.
+       \param periodic Whether each logical dimension is periodic.
+       \param partitioner The grid partitioner.
+    */
+    GlobalGrid( MPI_Comm comm,
+                const std::shared_ptr<GlobalMesh<MeshType>>& global_mesh,
+                const std::array<bool, num_space_dim>& periodic,
+                const BlockPartitioner<num_space_dim>& partitioner );
+    // const PartitionerType& partitioner );
+
+    //! \brief Get the owned number of cells in a given dimension of this
+    //! block. \param dim Spatial dimension.
     int ownedNumCell( const int dim ) const;
 
     //! \brief Get the global offset in a given dimension. This is where our
@@ -170,15 +212,51 @@ class GlobalGrid
                               const std::array<int, num_space_dim>& offset );
 
   private:
-    MPI_Comm _cart_comm;
-    std::shared_ptr<GlobalMesh<MeshType>> _global_mesh;
-    std::array<bool, num_space_dim> _periodic;
-    std::array<int, num_space_dim> _ranks_per_dim;
-    std::array<int, num_space_dim> _cart_rank;
     std::array<int, num_space_dim> _owned_num_cell;
     std::array<int, num_space_dim> _global_cell_offset;
-    std::array<bool, num_space_dim> _boundary_lo;
-    std::array<bool, num_space_dim> _boundary_hi;
+    // const BlockPartitioner<num_space_dim> _partitioner;
+    // std::shared_ptr<PartitionerType> _partitioner;
+};
+
+//---------------------------------------------------------------------------//
+/*!
+  \brief Global logical grid, specialization for sparse grids
+  \tparam Scalar Mesh floating point type: default = 4
+  \tparam NumSpaceDim Spatial dimension: default = 3
+*/
+template <typename Device, class Scalar, unsigned long long CellPerTileDim = 4,
+          std::size_t NumSpaceDim = 3>
+class GlobalSparseGrid : GlobalGridBase<SparseMesh<Scalar, NumSpaceDim>>
+{
+  public:
+    //! Mesh type.
+    using mesh_type = SparseMesh<Scalar, NumSpaceDim>;
+
+    //! Spatial dimension
+    using GlobalGridBase<mesh_type>::num_space_dim;
+
+    //! Kokkos device type.
+    using device_type = Device;
+
+    //! Number of bits (per dimension) needed to index the cells inside a tile
+    static constexpr unsigned long long cell_bits_per_tile_dim =
+        bitCount( CellPerTileDim );
+
+    //! Number of cells inside each tile (per dimension), tile size reset to
+    //! power of 2
+    static constexpr unsigned long long cell_num_per_tile_dim =
+        1 << cell_bits_per_tile_dim;
+
+    GlobalSparseGrid( MPI_Comm comm,
+                      const std::shared_ptr<GlobalMesh<mesh_type>>& global_mesh,
+                      const std::array<bool, num_space_dim>& periodic,
+                      std::shared_ptr<SparseDimPartitioner<
+                          device_type, cell_num_per_tile_dim>>& partitioner );
+
+  private:
+    std::array<int, num_space_dim> _owned_num_cell;
+    std::array<int, num_space_dim> _global_cell_offset;
+    SparseDimPartitioner<device_type, cell_num_per_tile_dim> _partitioner;
 };
 
 //---------------------------------------------------------------------------//
@@ -191,6 +269,28 @@ class GlobalGrid
   \param periodic Whether each logical dimension is periodic.
   \param partitioner The grid partitioner.
 */
+// template <class MeshType, class PartitionerType>
+// std::shared_ptr<GlobalGrid<MeshType, PartitionerType>> createGlobalGrid(
+//     MPI_Comm comm, const std::shared_ptr<GlobalMesh<MeshType>>& global_mesh,
+//     const std::array<bool, MeshType::num_space_dim>& periodic,
+//     const DimBlockPartitioner<MeshType::num_space_dim>& partitioner )
+// {
+//     return std::make_shared<
+//         GlobalGrid<MeshType, DimBlockPartitioner<MeshType::num_space_dim>>>(
+//         comm, global_mesh, periodic, partitioner );
+// }
+
+// template <class MeshType, class PartitionerType>
+// std::shared_ptr<GlobalGrid<MeshType, PartitionerType>> createGlobalGrid(
+//     MPI_Comm comm, const std::shared_ptr<GlobalMesh<MeshType>>& global_mesh,
+//     const std::array<bool, MeshType::num_space_dim>& periodic,
+//     const ManualBlockPartitioner<MeshType::num_space_dim>& partitioner )
+// {
+//     return std::make_shared<
+//         GlobalGrid<MeshType,
+//         ManualBlockPartitioner<MeshType::num_space_dim>>>( comm, global_mesh,
+//         periodic, partitioner );
+// }
 template <class MeshType>
 std::shared_ptr<GlobalGrid<MeshType>>
 createGlobalGrid( MPI_Comm comm,
@@ -202,7 +302,7 @@ createGlobalGrid( MPI_Comm comm,
                                                    partitioner );
 }
 
-//---------------------------------------------------------------------------//
+//---------------------------------------------------------------------------/
 
 } // end namespace Cajita
 
@@ -212,6 +312,6 @@ createGlobalGrid( MPI_Comm comm,
 
 #include <Cajita_GlobalGrid_impl.hpp>
 
-//---------------------------------------------------------------------------//
+//---------------------------------------------------------------------------//u
 
-#endif // end CAJITA_GLOBALGRID_HPP
+#endif // !CAJITA_GLOBALGRID_NEW_HPP
